Fix initial health being a float

+++ engine/src/main/battlecode/world/GameMapIO.java
@@ -362,7 +362,7 @@ public final strictfp class GameMapIO {
                 int bodyID = bodyTable.robotIDs(i);
                 int bodyX = locs.xs(i);
                 int bodyY = locs.ys(i);
-                float health = bodyTable.healths(i);
+                int health = bodyTable.healths(i);
                 Team bodyTeam = TeamMapping.team(bodyTable.teamIDs(i));
                 initialBodies.add(new RobotInfo(bodyID, bodyTeam, bodyType, health, new MapLocation(bodyX, bodyY)));
             }
